fix(calc) : Change Data Type(int => float)

+++ src/main/java/Calculator.java
@@ -3,19 +3,19 @@ public class Calculator {
 
     }
 
-    public int add(int num1, int num2) {
+    public float add(float num1, float num2) {
         return num1 + num2;
     }
 
-    public int sub(int num1, int num2) {
+    public float sub(float num1, float num2) {
         return num1 - num2;
     }
 
-    public int mul(int num1, int num2) {
+    public float mul(float num1, float num2) {
         return num1 * num2;
     }
 
-    public float div(int num1, int num2) {
+    public float div(float num1, float num2) {
         return num1 / num2;
     }
